Handling rendering when AttachPoints change
Mizuhiki now handles the case when there is no AttachPoint set or the
attach point is changed.

Also the dom node is saves to the property domNode on the object to
match how Dojo does it, if we are actually rendering the whole control
and not just a part of it.

The TemplatedObject now has a startup function to mimic Dojo, that sets
the _started property. This should be able to be overridden when
inheriting from this. __super__ could  be used to call the inherited
object and call its startup, if we don't want to write the code again.

Should probably be a short for calling same function in the closest
inheritance, something like __superf__() maybe?